@@ -29,7 +29,7 @@ for path in sorted(src.rglob("*.py")):
 
     with mkdocs_gen_files.open(full_doc_path, "w") as fd:
         ident = ".".join(parts)
-        fd.write(f"::: {ident}")
+        fd.write(f"---\ntitle: {ident}\n---\n\n::: {ident}")
 
     mkdocs_gen_files.set_edit_path(full_doc_path, ".." / path.relative_to(root))
 
